pyproject.toml
# SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
# SPDX-License-Identifier: Apache-2.0

# Build
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
artifacts = ["aiperf/operator/ui/**"]

[project]
name = "aiperf"
version = "0.6.0"
description = "AIPerf is a package for performance testing of AI models"
readme = "README.md"
license = {text = "Apache-2.0"}
license-files = ["LICENSE", "ATTRIBUTIONS.md"]
authors = [
    { name = "NVIDIA Inc.", email = "sw-dl-dynamo@nvidia.com" },
]
requires-python = ">=3.10"
dependencies = [
  "aiofiles~=24.1.0",
  "async-timeout>=4.0.0; python_version < '3.11'",
  "aiohttp~=3.13.3",
  "cyclopts>=4,<5",
  "fastapi>=0.115,<1",
  "ffmpeg-python~=0.2.0",
  "datasets>=3.0",
  "huggingface-hub>=0.34.0,<2.0",
  "jinja2~=3.1.5", # NOTE: Versions prior to 3.1.5 have vuln exploits
  "jmespath~=1.0.1",
  "kaleido~=1.2.0",
  "matplotlib>=3.10.0",
  "msgspec>=0.19.0,<1.0.0",
  "numpy>=1.26.4,<3",
  "nvidia-ml-py", # Note: No version specified to be most compatible with CUDA version
  "orjson~=3.11.7",
  "pandas~=2.3.3",
  "pillow~=12.1.1",
  "plotly~=6.4.0",
  "dash~=3.1.0",
  "dash-bootstrap-components~=2.0.0",
  "prometheus_client~=0.23.1",
  "psutil~=7.0.0",
  "pyarrow>=18.0.0",
  "pydantic>=2.10.0,<3.0.0",
  "pydantic-settings>=2.10.0,<3.0.0",
  "pyzmq~=26.4.0",
  "rich~=14.1.0",
  "ruamel.yaml~=0.18.12",
  "seaborn~=0.13.2",
  "setproctitle~=1.3.6",
  "soundfile~=0.13.1",
  "starlette-compress>=1.0.1",
  "textual~=5.3.0",
  "tqdm>=4.67.1",
  "transformers>=4.56.0", # Lowest compatible version for dynamo backends
  "uvicorn[standard]>=0.34,<1",
  "uvloop~=0.21.0; platform_system != 'Windows'",
  "kopf>=1.42.0",
  "zstandard>=0.25.0",
  "python-multipart>=0.0.22",
  "kr8s>=0.20.15",
  "duckdb>=1.5.0",
]

# CLI Entrypoints
[project.scripts]
aiperf = "aiperf.cli:app"

# Plugin entry point definition to discover plugins at runtime
[project.entry-points."aiperf.plugins"]
aiperf = "aiperf.plugin:plugins.yaml"

[project.optional-dependencies]
dev = [
  "black>=25.1.0",
  "httpx>=0.27.0",
  "jsonschema>=4.23.0",
  "looptime>=0.5",
  "pre-commit>=4.2.0",
  "pytest-asyncio",
  "pytest-cov",
  "pytest>=7.0.0",
  "pytest-xdist>=3.8.0",
  "ruff>=0.14.0,<0.15.0",
  "scipy>=1.13.0",
  "trustme>=1.0.0",
  "pympler>=1.1",
]

[tool.ruff]
# Base settings for ruff
line-length = 88
indent-width = 4
exclude = ["__pycache__", "build", "dist", ".venv", "venv", "notebooks/TCO_calculator.ipynb", "new_config", "demo.py", "validate_examples.py"]

[tool.ruff.lint]
select = [
    # pycodestyle (except E501)
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
# Ignore line length errors, ruff format will handle this but
# can have some lines that are slightly over due to the way formatting works
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
# sys.path manipulation before imports is required for direct script execution
"dev/kube.py" = ["E402"]

[tool.pytest.ini_options]
# Basic pytest configuration
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "integration: marks tests as integration tests",
    "performance: marks tests as performance tests (deselected by default)",
    "ffmpeg: marks tests that require ffmpeg to be installed (deselected by default)",
    "stress: marks tests as stress tests that generate high load (deselected by default)",
    "slow: marks tests as slow (>= 3s, run by default but can be skipped with -m 'not slow')",
    "component_integration: marks tests as component integration tests",
    "server_unit: marks tests as unit tests for the mock server",
    "benchmark: marks tests as resource benchmarks for K8s sizing (deselected by default)",
    "gpu: marks tests requiring a GPU Kubernetes cluster (deselected by default)",
    "vllm: marks tests requiring a vLLM server (deselected by default)",
    "dynamo: marks tests requiring a Dynamo deployment (deselected by default)",
    "k8s: marks tests requiring a Kubernetes cluster",
    "k8s_slow: marks k8s tests that are known to be slow",
    "trtllm: marks tests requiring a TRT-LLM server (deselected by default)",
    "sglang: marks tests requiring an SGLang server (deselected by default)",
    "fern: marks tests that validate Fern documentation (requires fern CLI)",
]
# Better console output
console_output_style = "progress"
verbosity_assertions = 2
# Show extra test summary info
# Deselect slow/infra test suites by default - opt in explicitly:
#   pytest -m k8s, pytest -m gpu, pytest -m performance, pytest -m integration, etc.
# IDE note: When running a specific test file/function from IDE, markers are often bypassed
addopts = "--strict-markers -m 'not performance and not ffmpeg and not stress and not statistical and not component_integration and not integration and not server_unit and not gpu and not vllm and not dynamo and not trtllm and not sglang and not k8s and not fern'"
# Filter out known warnings from third-party libraries and test infrastructure
filterwarnings = [
    "ignore::RuntimeWarning:looptime",
    "ignore:There is no current event loop:DeprecationWarning",
    "ignore:coroutine.*was never awaited:RuntimeWarning",
    "ignore:unclosed file:ResourceWarning",
]

[tool.codespell]
skip = "*.pyc,*build*,tests/unit/transports/test_aiohttp_sse.py,tests/integration/assets/canary_reference_inputs.json,src/aiperf/server_metrics/units.py,src/aiperf/api/static/dashboard.html,src/aiperf/operator/ui/vendor/*"
ignore-words-list = "timeslice,timeslices,aks,ot"
